minimum functionality now achieved by employing schemawizard methods as a module
